+++ nomad-rest-api/package.json
@@ -2,12 +2,12 @@
   "name": "nomad-rest-api",
   "version": "3.2.2",
   "description": "REST API back-end for NOMAD system",
-  "main": "app.js",
+  "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "env-cmd -f ../env/api-no-docker.env node app.js",
-    "dev": "env-cmd -f ../env/api-no-docker.env nodemon app.js",
-    "dev-docker": "nodemon app.js"
+    "test": "env-cmd -f ../env/backend-test.env vitest  --run --reporter verbose",
+    "start": "env-cmd -f ../env/api-no-docker.env node server.js",
+    "dev": "env-cmd -f ../env/api-no-docker.env nodemon server.js",
+    "dev-docker": "nodemon server.js"
   },
   "keywords": [],
   "author": "",
@@ -31,6 +31,8 @@
   },
   "devDependencies": {
     "env-cmd": "^10.1.0",
-    "nodemon": "^2.0.18"
+    "nodemon": "^2.0.18",
+    "supertest": "^6.3.3",
+    "vitest": "^0.26.2"
   }
 }
